+++ Tests/SuperwallKitTests/Config/Assignments/Expression Evaluator/ExpressionEvaluatorTests.swift
@@ -22,8 +22,7 @@ final class ExpressionEvaluatorTests: XCTestCase {
       fromRule: .stub()
         .setting(\.expression, to: nil)
         .setting(\.expressionJs, to: nil),
-      eventData: .stub(),
-      isPreemptive: false
+      eventData: .stub()
     )
     XCTAssertTrue(result.shouldFire)
   }
@@ -41,8 +40,7 @@ final class ExpressionEvaluatorTests: XCTestCase {
     let result = await evaluator.evaluateExpression(
       fromRule: .stub()
         .setting(\.expression, to: "user.a == \"b\""),
-      eventData: EventData(name: "ss", parameters: [:], createdAt: Date()),
-      isPreemptive: false
+      eventData: EventData(name: "ss", parameters: [:], createdAt: Date())
     )
     XCTAssertTrue(result.shouldFire)
     XCTAssertNil(result.unsavedOccurrence)
@@ -61,8 +59,7 @@ final class ExpressionEvaluatorTests: XCTestCase {
       fromRule: .stub()
         .setting(\.expression, to: "user.a == \"b\"")
         .setting(\.occurrence, to: occurrence),
-      eventData: EventData(name: "ss", parameters: [:], createdAt: Date()),
-      isPreemptive: false
+      eventData: EventData(name: "ss", parameters: [:], createdAt: Date())
     )
     XCTAssertTrue(result.shouldFire)
     XCTAssertEqual(result.unsavedOccurrence, occurrence)
@@ -79,8 +76,7 @@ final class ExpressionEvaluatorTests: XCTestCase {
     let result = await evaluator.evaluateExpression(
       fromRule: .stub()
         .setting(\.expression, to: "params.a == \"b\""),
-      eventData: EventData(name: "ss", parameters: ["a": "b"], createdAt: Date()),
-      isPreemptive: false
+      eventData: EventData(name: "ss", parameters: ["a": "b"], createdAt: Date())
     )
     XCTAssertTrue(result.shouldFire)
   }
@@ -96,8 +92,7 @@ final class ExpressionEvaluatorTests: XCTestCase {
     let result = await evaluator.evaluateExpression(
       fromRule: .stub()
         .setting(\.expression, to: "device.platform == \"iOS\""),
-      eventData: EventData(name: "ss", parameters: ["a": "b"], createdAt: Date()),
-      isPreemptive: false
+      eventData: EventData(name: "ss", parameters: ["a": "b"], createdAt: Date())
     )
     XCTAssertTrue(result.shouldFire)
   }
@@ -113,8 +108,7 @@ final class ExpressionEvaluatorTests: XCTestCase {
     let result = await evaluator.evaluateExpression(
         fromRule: .stub()
           .setting(\.expression, to: "device.platform == \"Android\""),
-      eventData: EventData(name: "ss", parameters: ["a": "b"], createdAt: Date()),
-        isPreemptive: false
+      eventData: EventData(name: "ss", parameters: ["a": "b"], createdAt: Date())
     )
     XCTAssertFalse(result.shouldFire)
   }
@@ -130,8 +124,7 @@ final class ExpressionEvaluatorTests: XCTestCase {
     let result = await evaluator.evaluateExpression(
       fromRule: .stub()
         .setting(\.expression, to: "a == \"b\""),
-      eventData: .stub(),
-      isPreemptive: false
+      eventData: .stub()
     )
     XCTAssertFalse(result.shouldFire)
   }
@@ -161,8 +154,7 @@ final class ExpressionEvaluatorTests: XCTestCase {
     let result = await evaluator.evaluateExpression(
       fromRule: .stub()
         .setting(\.expressionJs, to: "function superwallEvaluator(){ return true }; superwallEvaluator"),
-      eventData: .stub(),
-      isPreemptive: false
+      eventData: .stub()
     )
     XCTAssertTrue(result.shouldFire)
   }
@@ -176,8 +168,7 @@ final class ExpressionEvaluatorTests: XCTestCase {
     let result = await evaluator.evaluateExpression(
       fromRule: .stub()
         .setting(\.expressionJs, to: "function superwallEvaluator(values) { return values.params.a ==\"b\" }; superwallEvaluator"),
-      eventData: EventData(name: "ss", parameters: ["a": "b"], createdAt: Date()),
-      isPreemptive: false
+      eventData: EventData(name: "ss", parameters: ["a": "b"], createdAt: Date())
     )
     XCTAssertTrue(result.shouldFire)
   }
@@ -191,8 +182,7 @@ final class ExpressionEvaluatorTests: XCTestCase {
     let result = await evaluator.evaluateExpression(
       fromRule: .stub()
         .setting(\.expressionJs, to: "function superwallEvaluator(values) { return values.params.a ==\"b\" }; superwallEvaluator"),
-      eventData: EventData(name: "ss", parameters: ["a": "b"], createdAt: Date()),
-      isPreemptive: false
+      eventData: EventData(name: "ss", parameters: ["a": "b"], createdAt: Date())
     )
     XCTAssertTrue(result.shouldFire)
   }
@@ -206,8 +196,7 @@ final class ExpressionEvaluatorTests: XCTestCase {
     let result = await evaluator.evaluateExpression(
       fromRule: .stub()
         .setting(\.expressionJs, to: "function superwallEvaluator(values) { return 1 == 1 }; superwallEvaluator"),
-      eventData: .stub(),
-      isPreemptive: false
+      eventData: .stub()
     )
     XCTAssertTrue(result.shouldFire)
   }
@@ -235,8 +224,7 @@ final class ExpressionEvaluatorTests: XCTestCase {
     let result = await evaluator.evaluateExpression(
       fromRule: .stub()
         .setting(\.expressionJs, to: ""),
-      eventData: .stub(),
-      isPreemptive: false
+      eventData: .stub()
     )
     XCTAssertFalse(result.shouldFire)
   }