+++ Tests/SuperwallKitTests/Paywall/Presentation/Internal Presentation/Operators/GetPaywallVcOperatorTests.swift
@@ -14,8 +14,6 @@ final class GetPaywallVcOperatorTests: XCTestCase {
 
   @MainActor
   func test_getPaywallViewController_error_userSubscribed() async {
-    let experiment = Experiment(id: "", groupId: "", variant: .init(id: "", type: .treatment, paywallId: ""))
-
     let statePublisher = PassthroughSubject<PaywallState, Never>()
     let stateExpectation = expectation(description: "Output a state")
     stateExpectation.expectedFulfillmentCount = 2
@@ -55,7 +53,7 @@ final class GetPaywallVcOperatorTests: XCTestCase {
     do {
       _ = try await Superwall.shared.getPaywallViewController(
         request: request,
-        experiment: experiment,
+        rulesOutcome: .init(triggerResult: .paywall(.init(id: "", groupId: "", variant: .init(id: "", type: .treatment, paywallId: "")))),
         debugInfo: [:],
         paywallStatePublisher: statePublisher,
         dependencyContainer: dependencyContainer
@@ -74,8 +72,6 @@ final class GetPaywallVcOperatorTests: XCTestCase {
 
   @MainActor
   func test_getPaywallViewController_error_userNotSubscribed() async {
-    let experiment = Experiment(id: "", groupId: "", variant: .init(id: "", type: .treatment, paywallId: ""))
-
     let statePublisher = PassthroughSubject<PaywallState, Never>()
     let stateExpectation = expectation(description: "Output a state")
     stateExpectation.expectedFulfillmentCount = 2
@@ -110,7 +106,7 @@ final class GetPaywallVcOperatorTests: XCTestCase {
     do {
       _ = try await Superwall.shared.getPaywallViewController(
         request: request,
-        experiment: experiment,
+        rulesOutcome: .init(triggerResult: .paywall(.init(id: "", groupId: "", variant: .init(id: "", type: .treatment, paywallId: "")))),
         debugInfo: [:],
         paywallStatePublisher: statePublisher,
         dependencyContainer: dependencyContainer
@@ -129,8 +125,6 @@ final class GetPaywallVcOperatorTests: XCTestCase {
 
   @MainActor
   func test_getPaywallViewController_success_paywallNotAlreadyPresented() async {
-    let experiment = Experiment(id: "", groupId: "", variant: .init(id: "", type: .treatment, paywallId: ""))
-
     let statePublisher = PassthroughSubject<PaywallState, Never>()
     let stateExpectation = expectation(description: "Output a state")
     stateExpectation.isInverted = true
@@ -154,7 +148,7 @@ final class GetPaywallVcOperatorTests: XCTestCase {
     do {
       _ = try await Superwall.shared.getPaywallViewController(
         request: request,
-        experiment: experiment,
+        rulesOutcome: .init(triggerResult: .paywall(.init(id: "", groupId: "", variant: .init(id: "", type: .treatment, paywallId: "")))),
         debugInfo: [:],
         paywallStatePublisher: statePublisher,
         dependencyContainer: dependencyContainer